deps/v8/test/inspector/debugger/reuse-locals-blocklists-not-inside-function-expected.txt

Test for Debugger.evaluateOnCallFrame and locals blocklist reuse

Running test: testContextChainNotInsideFunction
{
    className : Function
    description : class C {}
    objectId : <objectId>
    type : function
}

Running test: testContextChainInsideFunction
{
    className : Function
    description : class C {}
    objectId : <objectId>
    type : function
}
